# fix: optimise add index performance