Slice recs into multiple smaller chunks when invoking AddIndex from boostd. Rpc data transfer has proven to take the longest time out of all indexing process.

Example for 56mil indexes:
```
Index Generation ~ 6 minutes (~56.5 mil)
Inserting Index into DB ~ 15 minutes
Breakdown of index insertion:
       dataTransfer ~ 8 minutes
       addMultihashesToPieces = 2 minutes
       addPieceInfos = 5m30s
       createPieceMetadata = 1s
```

Local performance test results against 1mil index records
```
1. 8 goroutines for data tranfer, 4 for insertion - 17 sec
2. 1 goroutine for data transfer, 1 for insertion - 50 sec 
3. 1 goroutine for data transfer, 32 for insertion - 27 sec (current main)
4. 8 coroutines for data transfer, 32 for insertion - 1.5min
```

Hence pick option 1 as reasonable defaults