@nonsense 2023-11-03
@ischasny please post a comparison, before/after this PR, from the description it is not clear what the improvement is.